+++ nautilus_trader/backtest/node.py
@@ -179,7 +179,6 @@ class BacktestNode:
                 book_type=BookTypeParser.from_str_py(config.book_type),
                 routing=config.routing,
                 frozen_account=config.frozen_account,
-                bar_execution=config.bar_execution,
                 reject_stop_orders=config.reject_stop_orders,
             )